Move classes to more appropriate package
